Small change to code eval and send() to server
New function for javascript, send()
Whatever the send is called with, is sent to the server.

send("this.player.x=100");

Will move the player to x:100 position for example.

Also disabled "close connection on error" functionality, previously if
the client sent data that resulted in an error, the connection was
immediately closed. Not anymore.